defconfig: Disable CONFIG_PFT
as it is unsupported Disable the drivers/platform/msm/pft.c driver as it is unsupported by Qualcomm, and opens us up to a wide range of potential attack surfaces that has not been audited by anyone.

Qualcomm recommends that it be disabled, as it hooks into SELinux in some "interesting" ways, and the userspace portion of the code is not even part of the image.

Bug: 28588434
Change-Id: I7fb1258839db984b34074bb53a0174d63393d728
Qualcomm-case-number: 02440206